requirements/base.txt
# PROJECT DEPENDENCIES (requirements.txt)
# This file lists all the third-party Python libraries needed to run the project.
# The '==' syntax pins the package to an exact, specific version.
# Core Django

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-colorfield==0.8.0

# Database
psycopg2-binary==2.9.7
Pillow==10.0.1

# Configuration
python-decouple==3.8

# Async tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2